need to debug the saving
attempted to save invitations using documents.save() but the starting_letter wouldn't open using document = Documnent(file_path). Attempted  converting the .docx to .txt because I was able to open the starting_letter with the 'with open()' but now it says 'failed to open the file as file is not a zip file'. 

Any idea how else I can try and get it to save the invitations?